ai-demo/requirements.txt
flask
flask-cors
gunicorn
python-dotenv
mysql-connector-python
psycopg2-binary
